+++ src/exactly_lib/impls/types/files_source/impl/file_makers/utils.py
@@ -1,19 +1,22 @@
 from typing import Callable
 
 from exactly_lib.common.report_rendering.parts import failure_details as _fd_rendering
-from exactly_lib.common.report_rendering.text_doc import TextRenderer
 from exactly_lib.impls import file_properties
-from exactly_lib.impls.file_properties import FileType, PropertiesWithNegation
-from exactly_lib.impls.types.path import top_lvl_error_msg_rendering
+from exactly_lib.impls.file_properties import FileType, CheckResult
+from exactly_lib.impls.types.path import path_err_msgs
 from exactly_lib.test_case.hard_error import HardErrorException
 from exactly_lib.test_case.result import failure_details
 from exactly_lib.type_val_prims.described_path import DescribedPath
-from exactly_lib.util.render import combinators as rend_comb
 
 
 class NewFileCreator:
     _PATH_CREATION_FAILURE = 'Failed to create path'
-    _FILE_CHECK = file_properties.negation_of(file_properties.must_exist(follow_symlinks=True))
+    _FILE_EXISTENCE_CHECK = file_properties.must_exist(follow_symlinks=False)
+    _PROPERTIES_FOR_ERR_MSG = file_properties.PropertiesWithNegation(
+        True,
+        file_properties.new_properties_for_existence(follow_symlinks=False,
+                                                     file_exists=True),
+    )
 
     def __init__(self, maker: Callable[[DescribedPath], None]):
         self._maker = maker
@@ -24,20 +27,18 @@ class NewFileCreator:
             self._maker(path)
         except OSError as ex:
             failure = failure_details.FailureDetails(
-                rend_comb.SingletonSequenceR(
-                    top_lvl_error_msg_rendering.header_and_path_block(
-                        self._PATH_CREATION_FAILURE,
-                        path)
-                ),
+                path_err_msgs.line_header__primitive__path(
+                    self._PATH_CREATION_FAILURE,
+                    path),
                 ex
             )
             raise HardErrorException(_fd_rendering.FailureDetailsRenderer(failure))
 
     def _assert_is_valid_path(self, path: DescribedPath):
-        result = self._FILE_CHECK.apply(path.primitive)
-        if not result.is_success:
+        result = self._FILE_EXISTENCE_CHECK.apply(path.primitive)
+        if result.is_success:
             raise HardErrorException(
-                file_properties.render_failure__d(result.cause, path)
+                file_properties.render_failure__d(self._PROPERTIES_FOR_ERR_MSG, path)
             )
 
 
@@ -55,7 +56,11 @@ class ExistingFileModifier:
 
     def _assert_is_valid_path(self, path: DescribedPath):
         result = self._file_check.apply(path.primitive)
-        if not result.is_success:
-            raise HardErrorException(
-                file_properties.render_failure__d(result.cause, path)
-            )
+        _raise_hard_error_if_not_success(result, path)
+
+
+def _raise_hard_error_if_not_success(result: CheckResult, path: DescribedPath):
+    if not result.is_success:
+        raise HardErrorException(
+            file_properties.render_failure__d(result.cause, path)
+        )