+++ src/exactly_lib/type_val_deps/validation_step.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+
+class ValidationStep(Enum):
+    PRE_SDS = 1
+    POST_SDS = 2