+++ CHANGELOG.md
@@ -11,6 +11,7 @@ with exception of a "0." prefix until version 1 is released.
 
  - Type `files-source` - Produces a set of files (for populating a directory)   
  - Instruction `dir` - Populate a dir with contents given by a `files-source`
+ - Instruction `file` - Append contents to existing file
 
 ### Changed